requirements.txt
black==23.7.0
grpcio-tools==1.56.0
grpcio==1.56.0
protobuf==4.23.4
pyarrow==12.0.1
pytest==7.4.0